+++ test/libsolidity/syntaxTests/pragma/invalid_range_conjunction_range_second.sol
@@ -1,0 +1,3 @@
+pragma solidity 0.8.17 0.4.1 - 0.6.2;
+// ----
+// ParserError 1684: (0-37): Invalid version pragma. Expected the start of a version number but instead found character '-'. Version number is invalid or the pragma is not terminated with a semicolon.