+++ test/libsolidity/syntaxTests/pragma/invalid_range_conjunction_two_ranges.sol
@@ -1,0 +1,3 @@
+pragma solidity 0.4.0 - 0.4.1 0.4.1 - 0.6.2;
+// ----
+// ParserError 1684: (0-44): Invalid version pragma. You can only combine version ranges using the || operator.