+++ Changelog.md
@@ -15,6 +15,7 @@ Compiler Features:
  * Optimizer: Added optimization rule ``and(shl(X, Y), shl(X, Z)) => shl(X, and(Y, Z))``.
  * SMTChecker: Support Eldarica as a Horn solver for the CHC engine when using the CLI option ``--model-checker-solvers eld``. The binary `eld` must be available in the system.
  * SMTChecker: Make ``z3`` the default solver for the BMC and CHC engines instead of all solvers.
+ * Parser: More detailed error messages about invalid version pragmas.
 
 
 Bugfixes: